requirements.txt
fastapi==0.110.1
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
uvicorn==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
